Move define statements to paths
