+++ src/scripts/apiClientResolver.js
@@ -1,0 +1,5 @@
+define("apiClientResolver", [], function () {
+    return function () {
+        return window.ApiClient;
+    };
+});
