+++ src/scripts/connectionManager.js
@@ -1,0 +1,3 @@
+define("connectionManager", [], function () {
+    return ConnectionManager;
+});
